Demuxer: Fix Ogg header shift UB, sample-rate validation, FLAC block desync
- Cast header bytes to uint32_t before <<24 in the Vorbis/Opus/Speex header
  parsers (a byte promoted to int overflows, which trips UBSan).
- Reject non-positive or implausibly large sample rates in the Vorbis and
  Speex parsers instead of only == 0 (a top-bit-set value read into the
  signed rate field would otherwise pass).
- Normalize the FLAC container position to each metadata block's end after
  processing it; the per-type parse/skip helpers could leave the position
  short of or past the boundary on malformed blocks, desyncing the header
  chain. Guard DemuxedStream's timestamp division against a zero rate.

+++ src/demuxer/ogg/OpusHeaderParser.cpp
@@ -57,7 +57,7 @@ bool OpusHeaderParser::parseOpusTags(const unsigned char* data, size_t size) {
         return false;
     }
     uint32_t vendor_length = data[offset] | (data[offset+1] << 8) | 
-                             (data[offset+2] << 16) | (data[offset+3] << 24);
+                             (data[offset+2] << 16) | (static_cast<uint32_t>(data[offset+3]) << 24);
     offset += 4;
     
     // Resource limit check
@@ -81,7 +81,7 @@ bool OpusHeaderParser::parseOpusTags(const unsigned char* data, size_t size) {
         return false;
     }
     uint32_t comment_count = data[offset] | (data[offset+1] << 8) | 
-                             (data[offset+2] << 16) | (data[offset+3] << 24);
+                             (data[offset+2] << 16) | (static_cast<uint32_t>(data[offset+3]) << 24);
     offset += 4;
     
     // Resource limit check
@@ -100,7 +100,7 @@ bool OpusHeaderParser::parseOpusTags(const unsigned char* data, size_t size) {
         }
         
         uint32_t comment_length = data[offset] | (data[offset+1] << 8) | 
-                                  (data[offset+2] << 16) | (data[offset+3] << 24);
+                                  (data[offset+2] << 16) | (static_cast<uint32_t>(data[offset+3]) << 24);
         offset += 4;
         
         // Resource limit check
